Format imm as signed decimal (as we used to do).

+++ src/common/types_frontend.rs
@@ -83,13 +83,14 @@ impl Const {
         }
     }
 
-    pub fn as_bits(&self) -> u64 {
+    pub fn as_bits(&self) -> i64 {
+        #[allow(clippy::unnecessary_cast)]
         match self {
-            Const::Int(i) => *i as u64,
-            Const::Long(i) => *i as u64,
-            Const::UInt(i) => *i as u64,
-            Const::ULong(i) => *i,
-            Const::Double(f) => f.to_bits(),
+            Const::Int(i) => *i as i64,
+            Const::Long(i) => *i as i64,
+            Const::UInt(i) => *i as i64,
+            Const::ULong(i) => *i as i64,
+            Const::Double(f) => f.to_bits() as i64,
         }
     }
